q3/requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.9
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
aiofiles==24.1.0
jinja2==3.1.4
requests==2.32.4
pillow==10.4.0 